move messagechans to node
