+++ cli/start.go
@@ -192,7 +192,7 @@ func MakeStartCommand() *cobra.Command {
 			}
 			// If the context has a messageChans defined, we pass along the relevant information.
 			// For now this is mostly useful for the CLI integration tests.
-			messageChans, ok := tryGetContextMessageChans(cmd.Context())
+			messageChans, ok := node.TryGetContextMessageChans(cmd.Context())
 			if ok && messageChans.APIURL != nil {
 				messageChans.APIURL <- n.APIURL
 				close(messageChans.APIURL)
